+++ infinigen_examples/generate_nature.py
@@ -644,10 +644,6 @@ def compose_nature(output_folder, scene_seed, **params):
         camera_rigs[0], "cube", offset=Vector(), size=30
     )
 
-    butil.constrain_object(
-        cube_emitter, "COPY_LOCATION", use_offset=True, target=camera_rigs[0]
-    )
-
     def leaf_particles():
         gen = weather.FallingParticles(
             leaves.LeafFactoryV2(randint(1e7)),
@@ -655,7 +651,7 @@ def compose_nature(output_folder, scene_seed, **params):
         )
         return gen(overhead_emitter)
 
-    p.run_stage("leaf_particles", leaf_particles, prereq="trees")
+    p.run_stage("leaf_particles", leaf_particles)
 
     def rain_particles():
         gen = weather.FallingParticles(